add isfollow flag to find user

+++ auth/src/controllers/follow/unFollowUser.controller.ts
@@ -14,9 +14,15 @@ export const unFollowHandler:UnFollowHandler = async (req,res,next)=>{
   if (!user) 
     return next(new NotFound('user not found'));
   
+  const sourceUser = await Users.findById(req.loggedUser.id);
+  if (!sourceUser) 
+    return next(new NotFound('user not found'));
+
   const unFollow = await Follow.findOneAndDelete({follower:req.loggedUser.id  , following:req.params.userId});
   if (unFollow) {
     user.followCount.followers--;
+    sourceUser.followCount.following--;
+    await sourceUser.save();
     await user.save();
   }
